requirements.txt
Flask==2.3.3
Flask-RESTful==0.3.10
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
psycopg2-binary==2.9.9
nltk==3.8.1
gunicorn==21.2.0
twilio==8.10.0
g4f==0.5.3.2
firebase-admin==6.5.0
requests==2.31.0
openai>=1.0.0
